Scrapper/requirements.txt
beautifulsoup4==4.12.2 
pandas==2.0.3
pyspark==3.4.1
requests==2.31.0
flake8==6.0.0
lxml==4.9.3
Unidecode==1.3.6
s3fs==2023.6.0
python-dotenv==1.0.0